+++ package.json
@@ -22,12 +22,15 @@
     "publish:linux": "yarn compile && electron-builder --x64 --linux --publish always"
   },
   "devDependencies": {
+    "@babel/core": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-react": "^7.12.1",
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
     "@types/webpack-env": "^1.15.3",
     "@typescript-eslint/eslint-plugin": "^4.22.1",
     "@typescript-eslint/parser": "^4.22.1",
-    "electron": "9.4.4",
+    "electron": "11.1.1",
     "electron-builder": "^22.9.1",
     "electron-webpack": "^2.8.2",
     "electron-webpack-ts": "^4.0.1",
@@ -45,18 +48,22 @@
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
-    "babel-cli": "6",
-    "babel-preset-react-app": "3",
     "builder-util-runtime": "^8.7.4",
     "electron-store": "^7.0.2",
     "electron-updater": "^4.3.8",
     "minimist": "^1.2.5",
     "pretty-bytes": "^5.6.0",
-    "react": "^17.0.1",
+    "react": "^17.0.2",
     "react-dom": "^17.0.1"
   },
   "bugs": {
     "url": "https://github.com/KadenBiel/DVD-TS/issues"
   },
-  "homepage": "https://github.com/KadenBiel/DVD-TS#readme"
+  "homepage": "https://github.com/KadenBiel/DVD-TS#readme",
+  "electronWebpack": {
+		"renderer": {
+			"webpackConfig": "webpack.renderer.config.js",
+			"webpackDllConfig": "webpack.renderer.config.js"
+		}
+	}
 }
