+++ src/common/ipc-messages.ts
@@ -13,7 +13,10 @@ export enum IpcRendererMessages {
 	RETURN_VERSION = 'RETURN_VERSION',
 	AUTO_UPDATER_STATE = 'AUTO_UPDATER_STATE',
 	SAVE_SETTINGS = 'SAVE_SETTINGS',
-	CLEAR_SETTINGS = 'CLEAR_SETTINGS'
+	CLEAR_SETTINGS = 'CLEAR_SETTINGS',
+	GET_SETTINGS = 'GET_SETTINGS',
+	START_DVD = 'START_DVD',
+	RETURN_SETTINGS = 'RETURN_SETTINGS'
 }
 
 export interface AutoUpdaterState {
