+++ src/renderer/Settings.tsx
@@ -1,30 +1,26 @@
 import React from 'react';
-//import TextField from '@material-ui/core/TextField';
 import makeStyles from '@material-ui/core/styles/makeStyles';
-//import withStyles from '@material-ui/core/styles/withStyles';
 import Box from '@material-ui/core/Box';
-//import FormControlLabel from '@material-ui/core/FormControlLabel';
-//import Radio from '@material-ui/core/Radio';
-//import Checkbox from '@material-ui/core/Checkbox';
-//import RadioGroup from '@material-ui/core/RadioGroup';
-//import MuiDivider from '@material-ui/core/Divider';
 import Typography from '@material-ui/core/Typography';
-//import Grid from '@material-ui/core/Grid';
 import ChevronLeft from '@material-ui/icons/ArrowBack';
 import IconButton from '@material-ui/core/IconButton';
-//import Alert from '@material-ui/lab/Alert';
-//import Slider from '@material-ui/core/Slider';
-//import Tooltip from '@material-ui/core/Tooltip';
-//import Dialog from '@material-ui/core/Dialog';
-//import DialogTitle from '@material-ui/core/DialogTitle';
-//import DialogContent from '@material-ui/core/DialogContent';
-//import DialogActions from '@material-ui/core/DialogActions';
-//import Button from '@material-ui/core/Button';
-//import DialogContentText from '@material-ui/core/DialogContentText';
+import { ipcRenderer } from 'electron';
+import { IpcRendererMessages } from '../common/ipc-messages';
 
 interface StyleInput {
 	open: boolean;
 }
+
+let dvdSpeed, size, colors, askUpdate;
+
+ipcRenderer.send(IpcRendererMessages.GET_SETTINGS)
+ipcRenderer.on(IpcRendererMessages.RETURN_SETTINGS, (event, settings) => {
+    ipcRenderer.removeAllListeners(IpcRendererMessages.RETURN_SETTINGS);
+    size = settings.size;
+	dvdSpeed = settings.dvdSpeed;
+    colors = settings.colors;
+	askUpdate = settings.askUpdate
+});
 
 const useStyles = makeStyles((theme) => ({
 	root: {
@@ -101,8 +97,8 @@ const useStyles = makeStyles((theme) => ({
 
 export interface SettingsProps {
 	open: boolean;
-	onClose: () => void;
-}
+	onClose: () => void
+};
 
 const Settings: React.FC<SettingsProps> = function ({ open, onClose }: SettingsProps) {
 	const classes = useStyles({ open });
@@ -114,6 +110,12 @@ const Settings: React.FC<SettingsProps> = function ({ open, onClose }: SettingsP
 					size="small"
 					onClick={() => {
 						onClose();
+						ipcRenderer.send(IpcRendererMessages.SAVE_SETTINGS, {
+							size: size,
+							dvdSpeed: dvdSpeed,
+							colors: colors,
+							askUpdate: askUpdate
+						})
 					}}
 				>
 					<ChevronLeft htmlColor="#777" />
