+++ tsconfig.json
@@ -2,7 +2,8 @@
     "extends": "./node_modules/electron-webpack/tsconfig-base.json",
     "compilerOptions": {
         "jsx": "react",
-        "skipLibCheck": true,
+        "skipLibCheck": false,
         "resolveJsonModule": true,
+        "strict": false
     },
 }
