+++ webpack.config.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+module.exports = {
+    resolve: {
+        alias: {
+            react: path.resolve(__dirname, 'node_modules', 'react')
+        }
+    }
+}
