+++ webpack.renderer.config.js
@@ -1,0 +1,5 @@
+module.exports = function(config) {
+    // ...
+    config.externals = [];
+    return config;
+}
